♻️ Extract OpenRouter client and message filtering helpers
Consolidates OpenRouter client creation to single shared module,
avoiding duplicate instantiation in route handler and concierge.

Extracts reasoning/thinking block filtering to reusable helper
function for multi-turn API compatibility.

Changes:
- lib/ai/openrouter.ts: Shared OpenRouter client with lazy init
- lib/ai/messages.ts: filterReasoningFromMessages() helper
- Updated route.ts and concierge/index.ts to use shared modules

Verified: All routing evals pass (25/25, 100% metrics)
